+++ swift_browser_ui_frontend/src/pages/SelectPage.vue
@@ -9,45 +9,34 @@
         v-for="project in projects.filter(pro => pro.tainted)"
         :key="project.id"
       >
-        <b-button
-          tag="a"
-          expanded
+        <c-button
           :href="'/lock/'.concat(project.id)"
-          size="is-large"
-          type="is-primary is-outlined"
+          size="large"
+          outlined
+          fit
         >
           {{ project.name }}
-        </b-button>
+        </c-button>
       </p>
       <p>
-        <b-button
-          tag="a"
-          expanded
+        <c-button
           href="/lock/none"
-          size="is-large"
-          type="is-primary is-outlined"
+          size="large"
+          outlined
+          fit
         >
           {{ $t("message.select.unrestricted" ) }}
-        </b-button>
+        </c-button>
       </p>
       <p>
-        <b-field class="locale-changed block center">
-          <b-select
-            v-model="$i18n.locale"
-            placeholder="Language"
-            icon="earth"
-            expanded
-            @input="setCookieLang ()"
-          >
-            <option
-              v-for="lang in langs"
-              :key="lang.value"
-              :value="lang.value"
-            >
-              {{ lang.ph }}
-            </option>
-          </b-select>
-        </b-field>
+        <c-select
+          v-csc-model="$i18n.locale"
+          class="locale-changed block center"
+          placeholder="Language"
+          :items="langs"
+          fit
+          @input="setCookieLang ()"
+        />
       </p>
     </div>
   </div>